requirements.txt
python-dotenv==1.1.0
heyoo==0.0.8
flask ==2.2.3
langchain== 0.0.136
llama_index==0.5.11